requirements-full.txt
astroid==2.2.5
bcrypt==3.1.6
cffi==1.12.2
Click==7.0
cmake==3.13.3
dlib==19.17.0
face-recognition==1.2.3
face-recognition-models==0.3.0
Flask==1.0.2
Flask-Bcrypt==0.7.1
isort==4.3.15
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.16.2
Pillow==5.4.1
pycparser==2.19
pylint==2.3.1
pymongo==3.7.2
schedule==0.6.0
six==1.12.0
typed-ast==1.3.1
Werkzeug==0.14.1
wrapt==1.11.1
